Move Composite912Codec under org.opensearch.index.codec.composite.composite912 package
