+++ server/src/main/java/org/opensearch/index/codec/composite/composite912/package-info.java
@@ -1,0 +1,12 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ *  Responsible for handling all composite index codecs and operations associated with Composite912 codec
+ */
+package org.opensearch.index.codec.composite.composite912;
