+++ server/src/main/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -1,2 +1,2 @@
 org.opensearch.index.codec.composite.composite99.Composite99Codec
-org.opensearch.index.codec.composite.Composite912Codec
+org.opensearch.index.codec.composite.composite912.Composite912Codec
