+++ server/src/test/java/org/opensearch/index/codec/CodecTests.java
@@ -48,7 +48,7 @@ import org.opensearch.common.settings.Settings;
 import org.opensearch.env.Environment;
 import org.opensearch.index.IndexSettings;
 import org.opensearch.index.analysis.IndexAnalyzers;
-import org.opensearch.index.codec.composite.Composite912Codec;
+import org.opensearch.index.codec.composite.composite912.Composite912Codec;
 import org.opensearch.index.engine.EngineConfig;
 import org.opensearch.index.mapper.MapperService;
 import org.opensearch.index.similarity.SimilarityService;
